# openapi.yaml
openapi: 3.0.1
info:
  title: portal NeoHoods API
  version: "2.0"
  description: |
    # Summary
    portal NeoHoods API
servers:
  - url: https://portal.neohoods.com
    description: portal neohoods API
paths:
...
  /admin/digital-locks:
    get:
      operationId: getDigitalLocks
      summary: Retrieve all digital locks
      tags:
        - digital-lock-admin-api
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: name
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/DigitalLockType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DigitalLockStatus"
      responses:
        "200":
          description: List of digital locks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DigitalLock"

    post:
      operationId: createDigitalLock
      summary: Create a new digital lock
      tags:
        - digital-lock-admin-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DigitalLock"
      responses:
        "201":
          description: Digital lock created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalLock"
        "400":
          $ref: "#/components/responses/BadRequest"
  /admin/digital-locks/{lock-id}:
    get:
      operationId: getDigitalLock
      summary: Retrieve a digital lock by ID
      tags:
        - digital-lock-admin-api
      parameters:
        - name: lock-id
          in: path
          required: true
          description: Digital lock ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Digital lock details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalLock"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateDigitalLock
      summary: Update a digital lock
      tags:
        - digital-lock-admin-api
      parameters:
        - name: lock-id
          in: path
          required: true
          description: Digital lock ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DigitalLock"
      responses:
        "200":
          description: Digital lock updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalLock"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"

    delete:
      operationId: deleteDigitalLock
      summary: Delete a digital lock
      tags:
        - digital-lock-admin-api
      parameters:
        - name: lock-id
          in: path
          required: true
          description: Digital lock ID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Digital lock deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/digital-locks/{lock-id}/status:
    patch:
      operationId: updateDigitalLockStatus
      summary: Update digital lock status
      tags:
        - digital-lock-admin-api
      parameters:
        - name: lock-id
          in: path
          required: true
          description: Digital lock ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: "#/components/schemas/DigitalLockStatus"
      responses:
        "200":
          description: Digital lock status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalLock"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
  /admin/digital-locks/{lock-id}/generate-code:
    post:
      operationId: generateDigitalLockAccessCode
      summary: Generate a temporary access code for a digital lock
      tags:
        - digital-lock-admin-api
      parameters:
        - name: lock-id
          in: path
          required: true
          description: Digital lock ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                duration:
                  type: integer
                  description: Duration in hours
                  example: 24
                reason:
                  type: string
                  description: Reason for generating the code
                  example: "Temporary access for maintenance"
      responses:
        "200":
          description: Access code generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessCode"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
...
  securitySchemes:
    BearerAuthOAuth:
      description: Bearer Token.
      type: http
      scheme: bearer
...
  schemas:
...
    AccessCode:
      type: object
      required:
        - code
        - generatedAt
        - expiresAt
      properties:
        code:
          type: string
          example: "ABC123"
        generatedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
          example: true
...
    CodedError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: A unique error code that identifies the type of error
          example: AUTH001
        message:
          type: string
          description: A human-readable error message
          example: The specified user could not be found
        traceId:
          type: string
          description: A unique identifier for tracing this error instance
          example: 550e8400-e29b-41d4-a716-446655440000
        documentationUrl:
          type: string
          description: URL to the documentation about this error
          example: https://docs.portal.neohoods.com/errors/AUTH001
        variables:
          type: object
          additionalProperties: true
          description: Additional context variables for the error
          example:
            field: "email"
            value: "invalid@email"
...
    DigitalLock:
      type: object
      required:
        - id
        - name
        - type
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Entrée principale"
        type:
          $ref: "#/components/schemas/DigitalLockType"
        status:
          $ref: "#/components/schemas/DigitalLockStatus"
        ttlockConfig:
          $ref: "#/components/schemas/TtlockConfig"
        nukiConfig:
          $ref: "#/components/schemas/NukiConfig"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TtlockConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          example: "ttlock-001"
        location:
          type: string
          example: "Entrée principale"
        batteryLevel:
          type: integer
          example: 85
        signalStrength:
          type: integer
          example: -45
        lastSeen:
          type: string
          format: date-time
    NukiConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          example: "nuki-001"
        token:
          type: string
          example: "nuki-token-789"
        batteryLevel:
          type: integer
          example: 95
        lastSeen:
          type: string
          format: date-time
    DigitalLockType:
      type: string
      enum:
        - TTLOCK
        - NUKI
        - YALE
      example: "TTLOCK"
    DigitalLockStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - ERROR
      example: "ACTIVE"
...
  responses:
...
    BadRequest:
      description: Bad Request - The request was invalid or cannot be served
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodedError"
          examples:
            validation_error:
              value:
                code: VAL001
                message: Validation failed
                variables:
                  fields:
                    - field: email
                      message: must be a valid email
            invalid_input:
              value:
                code: VAL002
                message: Missing required field
                variables:
                  field: username
...
    NotFound:
      description: Not Found - The requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodedError"
          example:
            code: RES001
            message: The requested resource was not found
...